screenshot-app/pyproject.toml
[project]
name = "screenshot-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=11.1.0",
    "pyautogui>=0.9.54",
    "streamlit>=1.44.1",
    "streamlit-drawable-canvas>=0.9.3",
]
